requirements.txt
asgiref==3.8.0
bitly-api-py3==0.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
idna==3.6
pyshorteners==1.0.1
requests==2.31.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
